# frontend/content-hub-frontend/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /person/details:
    get:
      tags:
      - person-controller
      operationId: getPersonDetails
      parameters:
      - name: personId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PersonResponseDto"
...
components:
  schemas:
...
    PersonCreditsCastDto:
      type: object
      properties:
        adult:
          type: boolean
        backdropPath:
          type: string
        genreIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        originalLanguage:
          type: string
        overview:
          type: string
        popularity:
          type: number
        posterPath:
          type: string
        voteAverage:
          type: number
        voteCount:
          type: integer
          format: int32
        character:
          type: string
        creditId:
          type: string
        originalTitle:
          type: string
        releaseDate:
          type: string
        releaseYear:
          type: string
        title:
          type: string
        episodeCount:
          type: integer
          format: int32
        mediaType:
          type: string
    PersonCreditsCrewDto:
      type: object
      properties:
        adult:
          type: boolean
        backdropPath:
          type: string
        genreIds:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: integer
          format: int32
        originalLanguage:
          type: string
        overview:
          type: string
        popularity:
          type: number
        posterPath:
          type: string
        voteAverage:
          type: number
        voteCount:
          type: integer
          format: int32
        creditId:
          type: string
        department:
          type: string
        job:
          type: string
        originalTitle:
          type: string
        releaseDate:
          type: string
        releaseYear:
          type: string
        title:
          type: string
        episodeCount:
          type: integer
          format: int32
        mediaType:
          type: string
    PersonResponseDto:
      type: object
      properties:
        adult:
          type: boolean
        alsoKnownAs:
          type: array
          items:
            type: string
        biography:
          type: string
        birthday:
          type: string
        deathday:
          type: string
        gender:
          type: integer
          format: int32
        homepage:
          type: string
        id:
          type: integer
          format: int32
        imdbId:
          type: string
        knownForDepartment:
          type: string
        name:
          type: string
        placeOfBirth:
          type: string
        popularity:
          type: number
        profilePath:
          type: string
        genderValue:
          type: string
        castCount:
          type: integer
          format: int32
        crewCount:
          type: integer
          format: int32
        cast:
          type: array
          items:
            $ref: "#/components/schemas/PersonCreditsCastDto"
        crew:
          type: array
          items:
            $ref: "#/components/schemas/PersonCreditsCrewDto"
...